Start implementation of main Zeus program
  + Fix signal handling in Dummies
  + Fix small bug in LCP

In Dummies, the actual dummy is run in a child process.
By killing the parent, the child gets orphaned but still
working. Now, we are intercepting the signals in the parent
and then forwarding to the child, killing the child before
the parent.

TODO: actually connect the Dummies responses to LCP

Changes to be committed:
	modified:   Dummies/
	modified:   LCP/
	new file:   Zeus.py