Connect RCP to Dummies via Zeus
The lcp_provider is being redundant, it is the main server program (in
this case, Zeus) that has a thread that connects to the dummy. In the
future, lcp_provider will be fully inside Zeus.py

Also, the responsability files hack can be safely erased.

Missing is the routing of RCP requests in Heimdall (lcp_bridge).
How can we know where are the dummies (Zeus or Anakin). Duplicate info?
If asking them, how do we know Anakin is not lying to create a
side-channel using RCP? Send to both?

Changes to be committed:
	modified:   Dummies/GenericDummy.py
	modified:   LCP/lcp_bridge.py
	modified:   LCP/lcp_provider.py
	modified:   Zeus.py