Add CP certificate exchange and validation in UP
  + Add pem to gitignore
  + Write small utils for certificate management (check and read)

The Architect is a mini-CA, and all certificates are generated at
runtime (except root cert/private key if available). Then, the others
must query the Architect (ARCHITECT_CP) for updates on CRLs and
certificates, and check them.

In this commit, the Architect generates a certificate for UP and revokes
it. Then, upon insertion of a newline on the CLI of Architect, it
generates a new, valid UP certificate. Zeus (ZEUS_UP) does not update
the dummy "brakes" until this valid certificate is available.

Changes to be committed:
	modified:   .gitignore
	new file:   Architect.py
	new file:   CP/cp_configuration.py
	new file:   CP/cp_server.py
	modified:   CP/cp_utils.py
	modified:   CP/create_architect_certs.py
	modified:   RCP/rcp_bridge.py
	modified:   UP/up_client.py
	modified:   UP/up_server.py
	modified:   Zeus.py